+++ packages/snap-controller/src/Search/SearchController.ts
@@ -164,6 +164,26 @@ export class SearchController extends AbstractController {
 			params.tracking.userId = userId;
 		}
 
+		if (!this.config.globals?.personalization?.disabled) {
+			const cartItems = this.tracker.getCartItems();
+			if (cartItems.length) {
+				params.personalization = params.personalization || {};
+				params.personalization.cart = cartItems.join(',');
+			}
+
+			const lastViewedItems = this.tracker.getLastViewedItems();
+			if (lastViewedItems.length) {
+				params.personalization = params.personalization || {};
+				params.personalization.lastViewed = lastViewedItems.join(',');
+			}
+
+			const shopperId = this.tracker.getShopperId()?.shopperId;
+			if (shopperId) {
+				params.personalization = params.personalization || {};
+				params.personalization.shopper = shopperId;
+			}
+		}
+
 		return params;
 	}