+++ packages/snap-preact/src/create/createSearchController.ts
@@ -1,3 +1,5 @@
+import { configure as configureMobx, extendObservable } from 'mobx';
+
 import { SearchController } from '@searchspring/snap-controller';
 import { Client } from '@searchspring/snap-client';
 import { SearchStore } from '@searchspring/snap-store-mobx';
@@ -6,8 +8,9 @@ import { EventManager } from '@searchspring/snap-event-manager';
 import { Profiler } from '@searchspring/snap-profiler';
 import { Logger } from '@searchspring/snap-logger';
 import { Tracker } from '@searchspring/snap-tracker';
-
 import type { SnapControllerServices, SnapSearchControllerConfig } from '../types';
+
+configureMobx({ useProxies: 'never' });
 
 export default (config: SnapSearchControllerConfig, services?: SnapControllerServices): SearchController => {
 	const urlManager = services?.urlManager || new UrlManager(new UrlTranslator(config.url), reactLinker);